perf(types): reduce instance inheritance overhead
