+++ types/instance-lifecycle.d.ts
@@ -1,10 +1,3 @@
-import {
-  CallbackFunc as LightMyRequestCallback,
-  Chain as LightMyRequestChain,
-  InjectOptions,
-  Response as LightMyRequestResponse
-} from 'light-my-request'
-
 export interface FastifyListenOptions {
   /** Default to `0` (picks the first available open port). */
   port?: number
@@ -25,17 +18,4 @@ export interface FastifyListenOptions {
   signal?: AbortSignal
   /** Resolves the text logged after the server starts. */
   listenTextResolver?: (address: string) => string
-}
-
-/** Injection and listening members of a Fastify instance. */
-export interface FastifyInstanceLifecycle {
-  inject(opts: InjectOptions | string, cb: LightMyRequestCallback): void
-  inject(opts: InjectOptions | string): Promise<LightMyRequestResponse>
-  inject(): LightMyRequestChain
-  listen(
-    opts: FastifyListenOptions,
-    callback: (err: Error | null, address: string) => void
-  ): void
-  listen(opts?: FastifyListenOptions): Promise<string>
-  listen(callback: (err: Error | null, address: string) => void): void
 }
