+++ types/instance.d.ts
@@ -1,31 +1,80 @@
-import { ConstructorAction, ProtoAction } from './content-type-parser'
+import { AddressInfo } from 'node:net'
+import { ConstraintStrategy } from 'find-my-way'
+import {
+  CallbackFunc as LightMyRequestCallback,
+  Chain as LightMyRequestChain,
+  InjectOptions,
+  Response as LightMyRequestResponse
+} from 'light-my-request'
+import {
+  AddContentTypeParser,
+  ConstructorAction,
+  FastifyBodyParser,
+  ProtoAction,
+  getDefaultJsonParser,
+  hasContentTypeParser,
+  removeAllContentTypeParsers,
+  removeContentTypeParser
+} from './content-type-parser'
 import { preHandlerAsyncHookHandler, preHandlerHookHandler, preValidationAsyncHookHandler, preValidationHookHandler } from './hooks'
 import { FastifyBaseLogger, FastifyChildLoggerFactory } from './logger'
-import { FastifyInstanceCore } from './instance-core'
-import { FastifyInstanceDecorators } from './instance-decorators'
 import { FastifyInstanceHooks } from './instance-hooks'
-import { FastifyInstanceLifecycle } from './instance-lifecycle'
-import { FastifyInstanceRouting } from './instance-routing'
-import { FastifyInstanceSchema } from './instance-schema'
+import { FastifyListenOptions } from './instance-lifecycle'
+import { FindMyWayFindResult, PrintRoutesOptions } from './instance-routing'
 import { FastifyRegister } from './register'
 import { FastifyReply, FastifyReplyForRoute } from './reply'
 import { FastifyRequest, FastifyRequestForRoute } from './request'
-import { FastifyRouterOptions } from './router-options'
-import { RouteGenericInterface, RouteHandlerMethod } from './route'
-import { FastifySchema } from './schema'
+import { FastifyRouterOptions, FindMyWayVersion } from './router-options'
+import { RouteGenericInterface, RouteHandlerMethod, RouteOptions, RouteShorthandMethod } from './route'
+import {
+  FastifySchema,
+  FastifySchemaCompiler,
+  FastifySchemaControllerOptions,
+  FastifySerializerCompiler,
+  SchemaErrorFormatter
+} from './schema'
 import {
   FastifyTypeProvider,
   FastifyTypeProviderDefault,
   SafePromiseLike
 } from './type-provider'
-import { FastifyTypeContextOf } from './type-context'
-import { ContextConfigDefault, RawReplyDefaultExpression, RawRequestDefaultExpression, RawServerBase, RawServerDefault } from './utils'
+import {
+  ContextConfigDefault,
+  RawReplyDefaultExpression,
+  RawRequestDefaultExpression,
+  RawServerBase,
+  RawServerDefault
+} from './utils'
 
 export type { FindMyWayVersion } from './router-options'
 export type { FindMyWayFindResult, PrintRoutesOptions } from './instance-routing'
 export type { FastifyListenOptions } from './instance-lifecycle'
 
 type ErrorHandlerResult = any | Promise<any>
+type AnyFunction = (...args: never[]) => unknown
+
+type GetterSetter<This, Value> = Value | {
+  getter: (this: This) => Value,
+  setter?: (this: This, value: Value) => void
+}
+
+type DecorationMethod<This, Instance, Return = Instance> = {
+  <
+    // Keep the historical generic order: decorate<Foo>() must retain its meaning.
+    Value extends (Property extends keyof This ? This[Property] : unknown),
+    Property extends string | symbol = string | symbol
+  >(
+    property: Property,
+    value: GetterSetter<This, Value extends AnyFunction
+      ? (this: This, ...args: Parameters<Value>) => ReturnType<Value>
+      : Value
+    >,
+    dependencies?: string[]
+  ): Return
+
+  (property: string | symbol): Return
+  (property: string | symbol, value: null | undefined, dependencies: string[]): Return
+}
 
 /**
  * Fastify server instance. Returned by the core `fastify()` method.
@@ -36,28 +85,159 @@ export interface FastifyInstance<
   RawReply extends RawReplyDefaultExpression<RawServer> = RawReplyDefaultExpression<RawServer>,
   Logger extends FastifyBaseLogger = FastifyBaseLogger,
   TypeProvider extends FastifyTypeProvider = FastifyTypeProviderDefault
-> extends FastifyInstanceCore<
-      FastifyTypeContextOf<RawServer, RawRequest, RawReply, Logger, TypeProvider>,
-      FastifyInstance<RawServer, RawRequest, RawReply, Logger, TypeProvider>
-    >,
-  FastifyInstanceDecorators<FastifyInstance<RawServer, RawRequest, RawReply, Logger, TypeProvider>, FastifyRequest,
-    FastifyReply>,
-  FastifyInstanceHooks<FastifyInstance<RawServer, RawRequest, RawReply, Logger, TypeProvider>, RawServer, RawRequest,
-    RawReply, Logger, TypeProvider>,
-  FastifyInstanceLifecycle,
-  FastifyInstanceRouting<
-      FastifyTypeContextOf<RawServer, RawRequest, RawReply, Logger, TypeProvider>,
-      FastifyInstance<RawServer, RawRequest, RawReply, Logger, TypeProvider>
-    >,
-  FastifyInstanceSchema<
-      FastifyTypeContextOf<RawServer, RawRequest, RawReply, Logger, TypeProvider>,
-      FastifyInstance<RawServer, RawRequest, RawReply, Logger, TypeProvider>
-    > {
+> extends FastifyInstanceHooks<FastifyInstance<RawServer, RawRequest, RawReply, Logger, TypeProvider>, RawServer,
+    RawRequest, RawReply, Logger, TypeProvider> {
+  server: RawServer
+  pluginName: string
+  prefix: string
+  version: string
+  log: Logger
+  listeningOrigin: string
+  addresses(): AddressInfo[]
+
+  addSchema(schema: unknown): FastifyInstance<RawServer, RawRequest, RawReply, Logger, TypeProvider>
+  getSchema(schemaId: string): unknown
+  getSchemas(): Record<string, unknown>
+
+  after(): FastifyInstance<RawServer, RawRequest, RawReply, Logger, TypeProvider> & SafePromiseLike<undefined>
+  after(afterListener: (err: Error | null) => void): FastifyInstance<RawServer, RawRequest, RawReply, Logger,
+    TypeProvider>
+
+  close(): Promise<undefined>
+  close(closeListener: () => void): undefined
+
+  /** Alias for `close()`. */
+  // @ts-ignore - type only available for @types/node >=17 or TypeScript >=5.2
+  [Symbol.asyncDispose](): Promise<undefined>
+
+  ready(): FastifyInstance<RawServer, RawRequest, RawReply, Logger, TypeProvider> & SafePromiseLike<undefined>
+  ready(readyListener: (err: Error | null) => void | Promise<void>): FastifyInstance<RawServer, RawRequest, RawReply,
+    Logger, TypeProvider>
+
   withTypeProvider<Provider extends FastifyTypeProvider>(): FastifyInstance<RawServer, RawRequest, RawReply, Logger,
     Provider>;
 
   register: FastifyRegister<FastifyInstance<RawServer, RawRequest, RawReply, Logger,
     TypeProvider> & SafePromiseLike<undefined>>;
+
+  decorate: DecorationMethod<FastifyInstance<RawServer, RawRequest, RawReply, Logger, TypeProvider>,
+    FastifyInstance<RawServer, RawRequest, RawReply, Logger, TypeProvider>>
+  decorateRequest: DecorationMethod<FastifyRequest, FastifyInstance<RawServer, RawRequest, RawReply, Logger,
+    TypeProvider>>
+  decorateReply: DecorationMethod<FastifyReply, FastifyInstance<RawServer, RawRequest, RawReply, Logger, TypeProvider>>
+  getDecorator<Value>(name: string | symbol): Value
+  hasDecorator(decorator: string | symbol): boolean
+  hasRequestDecorator(decorator: string | symbol): boolean
+  hasReplyDecorator(decorator: string | symbol): boolean
+  hasPlugin(name: string): boolean
+
+  addConstraintStrategy(strategy: ConstraintStrategy<FindMyWayVersion<RawServer>, unknown>): void
+  hasConstraintStrategy(strategyName: string): boolean
+  routing(req: RawRequest, res: RawReply): void
+
+  route<
+    RouteGeneric extends RouteGenericInterface = RouteGenericInterface,
+    ContextConfig = ContextConfigDefault,
+    const SchemaCompiler extends FastifySchema = FastifySchema
+  >(opts: RouteOptions<
+    RawServer,
+    RawRequest,
+    RawReply,
+    RouteGeneric,
+    ContextConfig,
+    SchemaCompiler,
+    TypeProvider,
+    Logger
+  >): FastifyInstance<RawServer, RawRequest, RawReply, Logger, TypeProvider>
+  delete: RouteShorthandMethod<RawServer, RawRequest, RawReply, TypeProvider, Logger>
+  get: RouteShorthandMethod<RawServer, RawRequest, RawReply, TypeProvider, Logger>
+  head: RouteShorthandMethod<RawServer, RawRequest, RawReply, TypeProvider, Logger>
+  patch: RouteShorthandMethod<RawServer, RawRequest, RawReply, TypeProvider, Logger>
+  post: RouteShorthandMethod<RawServer, RawRequest, RawReply, TypeProvider, Logger>
+  put: RouteShorthandMethod<RawServer, RawRequest, RawReply, TypeProvider, Logger>
+  options: RouteShorthandMethod<RawServer, RawRequest, RawReply, TypeProvider, Logger>
+  propfind: RouteShorthandMethod<RawServer, RawRequest, RawReply, TypeProvider, Logger>
+  proppatch: RouteShorthandMethod<RawServer, RawRequest, RawReply, TypeProvider, Logger>
+  mkcalendar: RouteShorthandMethod<RawServer, RawRequest, RawReply, TypeProvider, Logger>
+  mkcol: RouteShorthandMethod<RawServer, RawRequest, RawReply, TypeProvider, Logger>
+  copy: RouteShorthandMethod<RawServer, RawRequest, RawReply, TypeProvider, Logger>
+  move: RouteShorthandMethod<RawServer, RawRequest, RawReply, TypeProvider, Logger>
+  lock: RouteShorthandMethod<RawServer, RawRequest, RawReply, TypeProvider, Logger>
+  unlock: RouteShorthandMethod<RawServer, RawRequest, RawReply, TypeProvider, Logger>
+  trace: RouteShorthandMethod<RawServer, RawRequest, RawReply, TypeProvider, Logger>
+  report: RouteShorthandMethod<RawServer, RawRequest, RawReply, TypeProvider, Logger>
+  search: RouteShorthandMethod<RawServer, RawRequest, RawReply, TypeProvider, Logger>
+  query: RouteShorthandMethod<RawServer, RawRequest, RawReply, TypeProvider, Logger>
+  all: RouteShorthandMethod<RawServer, RawRequest, RawReply, TypeProvider, Logger>
+  hasRoute<
+    RouteGeneric extends RouteGenericInterface = RouteGenericInterface,
+    ContextConfig = ContextConfigDefault,
+    SchemaCompiler extends FastifySchema = FastifySchema
+  >(opts: Pick<RouteOptions<
+    RawServer,
+    RawRequest,
+    RawReply,
+    RouteGeneric,
+    ContextConfig,
+    SchemaCompiler,
+    TypeProvider
+  >, 'method' | 'url' | 'constraints'>): boolean
+  findRoute<
+    RouteGeneric extends RouteGenericInterface = RouteGenericInterface,
+    ContextConfig = ContextConfigDefault,
+    SchemaCompiler extends FastifySchema = FastifySchema
+  >(opts: Pick<RouteOptions<
+    RawServer,
+    RawRequest,
+    RawReply,
+    RouteGeneric,
+    ContextConfig,
+    SchemaCompiler,
+    TypeProvider
+  >, 'method' | 'url' | 'constraints'>): Omit<FindMyWayFindResult<RawServer>, 'store'>
+  supportedMethods: string[]
+  addHttpMethod(
+    method: string,
+    methodOptions?: { hasBody: boolean }
+  ): FastifyInstance<RawServer, RawRequest, RawReply, Logger, TypeProvider>
+  printRoutes(opts?: PrintRoutesOptions): string
+
+  inject(opts: InjectOptions | string, cb: LightMyRequestCallback): void
+  inject(opts: InjectOptions | string): Promise<LightMyRequestResponse>
+  inject(): LightMyRequestChain
+  listen(
+    opts: FastifyListenOptions,
+    callback: (err: Error | null, address: string) => void
+  ): void
+  listen(opts?: FastifyListenOptions): Promise<string>
+  listen(callback: (err: Error | null, address: string) => void): void
+
+  validatorCompiler: FastifySchemaCompiler<any> | undefined
+  setValidatorCompiler<Schema = FastifySchema>(
+    schemaCompiler: FastifySchemaCompiler<Schema>
+  ): FastifyInstance<RawServer, RawRequest, RawReply, Logger, TypeProvider>
+
+  serializerCompiler: FastifySerializerCompiler<any> | undefined
+  setSerializerCompiler<Schema = FastifySchema>(
+    schemaCompiler: FastifySerializerCompiler<Schema>
+  ): FastifyInstance<RawServer, RawRequest, RawReply, Logger, TypeProvider>
+
+  setSchemaController(
+    schemaControllerOpts: FastifySchemaControllerOptions
+  ): FastifyInstance<RawServer, RawRequest, RawReply, Logger, TypeProvider>
+  setReplySerializer(
+    replySerializer: (payload: unknown, statusCode: number) => string
+  ): FastifyInstance<RawServer, RawRequest, RawReply, Logger, TypeProvider>
+  setSchemaErrorFormatter(
+    errorFormatter: SchemaErrorFormatter
+  ): FastifyInstance<RawServer, RawRequest, RawReply, Logger, TypeProvider>
+
+  addContentTypeParser: AddContentTypeParser<RawServer, RawRequest, RouteGenericInterface, FastifySchema, TypeProvider>
+  hasContentTypeParser: hasContentTypeParser
+  removeContentTypeParser: removeContentTypeParser
+  removeAllContentTypeParsers: removeAllContentTypeParsers
+  getDefaultJsonParser: getDefaultJsonParser
+  defaultTextParser: FastifyBodyParser<string>
 
   /**
    * Set the 404 handler
